requirements.txt
discord.py>=2.3.0
anthropic>=0.40.0
APScheduler>=3.10.0
aiohttp>=3.9.0
aiofiles>=23.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
